+++ libnopegl/src/backends/gl/pipeline_gl.c
@@ -210,7 +210,8 @@ static int build_texture_bindings(struct pipeline *s, const struct pipeline_para
         const struct pipeline_texture_desc *texture_desc = &layout->textures_desc[i];
 
         if (texture_desc->type == NGLI_TYPE_IMAGE_2D ||
-            texture_desc->type == NGLI_TYPE_IMAGE_3D) {
+            texture_desc->type == NGLI_TYPE_IMAGE_3D ||
+            texture_desc->type == NGLI_TYPE_IMAGE_CUBE) {
             struct gpu_ctx_gl *gpu_ctx_gl = (struct gpu_ctx_gl *)s->gpu_ctx;
             struct glcontext *gl = gpu_ctx_gl->glcontext;
             const struct gpu_limits *limits = &gl->limits;
@@ -279,7 +280,8 @@ static void set_textures(struct pipeline *s, struct glcontext *gl)
         const struct texture_gl *texture_gl = (const struct texture_gl *)texture;
 
         if (texture_binding->desc.type == NGLI_TYPE_IMAGE_2D ||
-            texture_binding->desc.type == NGLI_TYPE_IMAGE_3D) {
+            texture_binding->desc.type == NGLI_TYPE_IMAGE_3D ||
+            texture_binding->desc.type == NGLI_TYPE_IMAGE_CUBE) {
             GLuint texture_id = 0;
             const GLenum access = get_gl_access(texture_binding->desc.access);
             GLenum internal_format = GL_RGBA8;
@@ -288,7 +290,8 @@ static void set_textures(struct pipeline *s, struct glcontext *gl)
                 internal_format = texture_gl->internal_format;
             }
             GLboolean layered = GL_FALSE;
-            if (texture_binding->desc.type == NGLI_TYPE_IMAGE_3D)
+            if (texture_binding->desc.type == NGLI_TYPE_IMAGE_3D ||
+                texture_binding->desc.type == NGLI_TYPE_IMAGE_CUBE)
                 layered = GL_TRUE;
             ngli_glBindImageTexture(gl, texture_binding->desc.binding, texture_id, 0, layered, 0, access, internal_format);
         } else {