+++ libnopegl/src/type.c
@@ -47,6 +47,7 @@ static const char * const type_map[NGLI_TYPE_NB] = {
     [NGLI_TYPE_SAMPLER_EXTERNAL_2D_Y2Y_EXT] = "__samplerExternal2DY2YEXT",
     [NGLI_TYPE_IMAGE_2D]                    = "image2D",
     [NGLI_TYPE_IMAGE_3D]                    = "image3D",
+    [NGLI_TYPE_IMAGE_CUBE]                  = "imageCube",
     [NGLI_TYPE_UNIFORM_BUFFER]              = "uniform",
     [NGLI_TYPE_STORAGE_BUFFER]              = "buffer",
 };